open_mpic_core_python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-mpic-core-python"
dynamic = ["version"]
description = 'MPIC Core library written in python.'
readme = "README.md"
requires-python = "==3.11.9"
license = "MIT"
keywords = []
authors = [
    { name = "Henry Birge-Lee", email = "henrybirgelee@gmail.com" },
    { name = "Grace Cimaszewski", email = "ghc2@princeton.edu"},
    { name = "Dmitry Sharkov", email = "dmitry.sharkov@sectigo.com" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "pyyaml==6.0.1",
    "requests>=2.32.3",
    "dnspython==2.6.1",
    "pydantic==2.8.2",
]

[project.optional-dependencies]
test = [
    "pytest==8.2.2",
    "pytest-cov",
    "pytest-mock==3.14.0",
    "pytest-html==4.1.1",
    "pytest-spec==3.2.0",
]

[project.urls]
Documentation = "https://github.com/open-mpic/open-mpic-core-python"
Issues = "https://github.com/open-mpic/open-mpic-core-python/issues"
Source = "https://github.com/open-mpic/open-mpic-core-python"

#[dirs.env]
#virtual = ".hatch"

[tool.hatch]
version.path = "src/open_mpic_core/__about__.py"
build.sources = ["src", "resources"]
build.targets.wheel.packages = ["src/open_mpic_core"]

[tool.hatch.envs.default]
skip-install = false
type="virtual"
path="venv"

[tool.hatch.envs.test]
features = [
    "test",
]

[tool.hatch.envs.test.scripts]
unit = "pytest"
unit-html = "pytest --html=testreports/index.html" # generate html report (warning: uses an aging plugin, 11-2023)
integration = "pytest tests/integration"
coverage = "pytest --cov=src/open_mpic_core --cov-report=term-missing --cov-report=html"

[tool.hatch.envs.hatch-test]
features = [
    "test",
    "provided"
]
default-args = ["tests/unit"]
randomize = true

[tool.pytest.ini_options]
pythonpath = [
    "src", "tests", "." # need root directory because it has some python utility files we are importing for integration tests
]
testpaths = [
    "tests/unit"
]
python_functions = [
    "*_should_*",
]
markers = [
    "integration: mark test as an integration test",
    "unit: mark test as a unit test",  # optional
]
addopts = [
    "--import-mode=prepend",  # explicit default, as the tests rely on it for proper import resolution
    "--spec"  # show test names in a more readable format in the console (warning: uses an aging plugin, 5-2021)
]
spec_header_format = "Spec for {test_case} ({path}):"
spec_test_format = "{result} {docstring_summary}"  # defaults to {name} if docstring is not present in test

[tool.coverage.run]
source = [
    "src/open_mpic_core"
]

[tool.coverage.report]
include_namespace_packages = true
omit = [
    "*/src/*/__about__.py",
]
